requirements.txt
Flask==2.3.3
spacy==3.7.2
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.23.5
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11
plotly==5.15.0
Werkzeug==2.3.7
